requirements.txt
Django==3.0.3
Pillow==7.0.0
django-phone-field==1.8.0
django-recaptcha==2.0.6